board-common: scan_coordinator poll-model refactor + display/camera fixes

+++ src/board_pipeline_display_lvgl.c
@@ -138,9 +138,10 @@ static bool push_frame_image_widget(lvgl_display_ctx_t *ctx,
         if (interval > interval_max) interval_max = interval;
         interval_count++;
         if (now - interval_last_log > 2000000) {
-            ESP_LOGI(TAG, "DISP INTERVAL: avg=%lld us  max=%lld us  n=%d",
-                     (long long)(interval_sum / interval_count),
-                     (long long)interval_max, interval_count);
+            /* %d (not %lld): nano-printf has no 64-bit conversion; us fit in int. */
+            ESP_LOGI(TAG, "DISP INTERVAL: avg=%d us  max=%d us  n=%d",
+                     (int)(interval_sum / interval_count),
+                     (int)interval_max, interval_count);
             interval_sum = 0;
             interval_max = 0;
             interval_count = 0;
